+++ packages/cli/lib/services/model.service.unit.test.ts
@@ -1,6 +1,8 @@
+import path from 'node:path';
 import { describe, expect, it } from 'vitest';
 import { buildModelsTS, fieldToTypescript, fieldsToTypescript } from './model.service.js';
 import type { NangoModel } from '@nangohq/types';
+import { load } from './config.service.js';
 
 function removeVersion(res: string) {
     return res.replace(/(v[0-9.]+)/, 'vTest');
@@ -64,6 +66,19 @@ describe('buildModelTs', () => {
             }
         });
         expect(removeVersion(res.split('\n').slice(0, 25).join('\n'))).toMatchSnapshot('');
+    });
+
+    it('should support all advanced syntax', () => {
+        const { response: parsed } = load(path.resolve(__dirname, `../../fixtures/nango-yaml/v2/advanced-syntax`));
+        const res = buildModelsTS({ parsed: parsed! });
+        const acc = [];
+        for (const line of res.split('\n')) {
+            if (line === '// ------ SDK') {
+                break;
+            }
+            acc.push(line);
+        }
+        expect(removeVersion(acc.join('\n'))).toMatchSnapshot();
     });
 });