+++ packages/orchestrator/lib/clients/validate.ts
@@ -27,7 +27,7 @@ export const syncArgsSchema = z.object({
 export const actionArgsSchema = z.object({
     type: z.literal('action'),
     actionName: z.string().min(1),
-    activityLogId: z.number().positive(),
+    activityLogId: z.union([z.number(), z.string()]),
     input: jsonSchema,
     ...commonSchemaArgsFields
 });
@@ -35,7 +35,7 @@ export const webhookArgsSchema = z.object({
     type: z.literal('webhook'),
     webhookName: z.string().min(1),
     parentSyncName: z.string().min(1),
-    activityLogId: z.number().positive(),
+    activityLogId: z.union([z.number(), z.string()]),
     input: jsonSchema,
     ...commonSchemaArgsFields
 });
@@ -43,7 +43,7 @@ export const postConnectionArgsSchema = z.object({
     type: z.literal('post-connection-script'),
     postConnectionName: z.string().min(1),
     fileLocation: z.string().min(1),
-    activityLogId: z.number().positive(),
+    activityLogId: z.union([z.number(), z.string()]),
     ...commonSchemaArgsFields
 });