+++ packages/server/lib/controllers/sync.controller.ts
@@ -352,7 +352,6 @@ class SyncController {
                 connection,
                 actionName: action_name,
                 input,
-                activityLogId,
                 environment_id: environmentId,
                 logCtx
             });