+++ packages/shared/lib/services/sync/manager.service.ts
@@ -132,9 +132,7 @@ export class SyncManagerService {
                 }
 
                 const createdSync = await createSync(connection.id as number, syncName);
-
-                // Is this not awaited on purpose?
-                void orchestrator.scheduleSyncHelper(
+                await orchestrator.scheduleSyncHelper(
                     connection,
                     createdSync as Sync,
                     syncConfig as ProviderConfig,